refactor: Make extensions use the FilemoonExtractor shared lib
