+++ multisrc/overrides/animestream/animenosub/additional.gradle
@@ -1,4 +1,5 @@
 dependencies {
+    implementation(project(':lib-filemoon-extractor'))
     implementation(project(':lib-streamsb-extractor'))
     implementation "dev.datlag.jsunpacker:jsunpacker:1.0.1"
 }
