+++ multisrc/overrides/animestream/animenosub/src/Animenosub.kt
@@ -2,12 +2,12 @@ package eu.kanade.tachiyomi.animeextension.en.animenosub
 
 import androidx.preference.ListPreference
 import androidx.preference.PreferenceScreen
-import eu.kanade.tachiyomi.animeextension.en.animenosub.extractors.FilemoonExtractor
 import eu.kanade.tachiyomi.animeextension.en.animenosub.extractors.StreamWishExtractor
 import eu.kanade.tachiyomi.animeextension.en.animenosub.extractors.VidMolyExtractor
 import eu.kanade.tachiyomi.animeextension.en.animenosub.extractors.VtubeExtractor
 import eu.kanade.tachiyomi.animeextension.en.animenosub.extractors.WolfstreamExtractor
 import eu.kanade.tachiyomi.animesource.model.Video
+import eu.kanade.tachiyomi.lib.filemoonextractor.FilemoonExtractor
 import eu.kanade.tachiyomi.lib.streamsbextractor.StreamSBExtractor
 import eu.kanade.tachiyomi.multisrc.animestream.AnimeStream
 import org.jsoup.nodes.Element
@@ -55,7 +55,7 @@ class Animenosub : AnimeStream(
                 WolfstreamExtractor(client).videosFromUrl(url, prefix)
             }
             url.contains("filemoon") -> {
-                FilemoonExtractor(client, headers).videosFromUrl(url, prefix)
+                FilemoonExtractor(client).videosFromUrl(url, prefix, headers)
             }
             else -> emptyList()
         }
