+++ multisrc/overrides/dooplay/cinemathek/additional.gradle
@@ -1,4 +1,5 @@
 dependencies {
+    implementation(project(':lib-filemoon-extractor'))
     implementation(project(':lib-dood-extractor'))
     implementation(project(':lib-streamlare-extractor'))
     implementation(project(':lib-streamsb-extractor'))
