+++ multisrc/overrides/dooplay/cinemathek/src/Cinemathek.kt
@@ -3,10 +3,10 @@ package eu.kanade.tachiyomi.animeextension.de.cinemathek
 import androidx.preference.ListPreference
 import androidx.preference.MultiSelectListPreference
 import androidx.preference.PreferenceScreen
-import eu.kanade.tachiyomi.animeextension.de.cinemathek.extractors.FilemoonExtractor
 import eu.kanade.tachiyomi.animeextension.de.cinemathek.extractors.StreamHideExtractor
 import eu.kanade.tachiyomi.animesource.model.Video
 import eu.kanade.tachiyomi.lib.doodextractor.DoodExtractor
+import eu.kanade.tachiyomi.lib.filemoonextractor.FilemoonExtractor
 import eu.kanade.tachiyomi.lib.streamlareextractor.StreamlareExtractor
 import eu.kanade.tachiyomi.lib.streamsbextractor.StreamSBExtractor
 import eu.kanade.tachiyomi.multisrc.dooplay.DooPlay
