+++ src/de/kool/build.gradle
@@ -11,6 +11,7 @@ ext {
 }
 
 dependencies {
+    implementation(project(':lib-filemoon-extractor'))
     implementation(project(':lib-streamtape-extractor'))
     implementation(project(':lib-voe-extractor'))
 }
