+++ src/en/nineanime/build.gradle
@@ -11,6 +11,7 @@ ext {
 }
 
 dependencies {
+    implementation(project(':lib-filemoon-extractor'))
     implementation(project(':lib-mp4upload-extractor'))
     implementation (project(':lib-streamtape-extractor'))
     implementation "dev.datlag.jsunpacker:jsunpacker:1.0.1"
